rename the library rename state rules remove useless metadata in state rules remove some redundancy
