+++ tests/tests.py
@@ -1,104 +1,92 @@
 import unittest
-from state_machine import states
+from state_machine.state_machine import transition, get_next_states, TransitionException
 
 
 class TestStateMachine(unittest.TestCase):
 
     def test_temperature_transition(self):
-        temperature_states = {
+        temperature_state_def = {
             'too hot': {
                 'next': ['just right'],
-                'prev': ['just right'],
             },
             'too cold': {
                 'next': ['just right'],
-                'prev': ['just right'],
             },
             'just right': {
                 'next': ['too hot', 'too cold'],
-                'prev': ['too hot', 'too cold'],
             }
         }
         room_state = 'just right'
-        room_state = states.transition(room_state, 'too hot', temperature_states)
+        room_state = transition(room_state, 'too hot', temperature_state_def)
         self.assertEqual(room_state, 'too hot')
-        room_state = states.transition(room_state, 'just right', temperature_states)
+        room_state = transition(room_state, 'just right', temperature_state_def)
         self.assertEqual(room_state, 'just right')
-        room_state = states.transition(room_state, 'too cold', temperature_states)
+        room_state = transition(room_state, 'too cold', temperature_state_def)
         self.assertEqual(room_state, 'too cold')
 
     def test_kanban_transition(self):
-        kanban_states = {
+        kanban_state_def = {
             'todo': {
                 'next': ['*'],
-                'prev': ['*'],
             },
             'doing': {
                 'next': ['*'],
-                'prev': ['*'],
             },
             'done': {
                 'next': ['*'],
-                'prev': ['*'],
             },
         }
         jira_issue_state = 'todo'
-        jira_issue_state = states.transition(jira_issue_state, 'doing', kanban_states)
+        jira_issue_state = transition(jira_issue_state, 'doing', kanban_state_def)
         self.assertEqual(jira_issue_state, 'doing')
-        jira_issue_state = states.transition(jira_issue_state, 'done', kanban_states)
+        jira_issue_state = transition(jira_issue_state, 'done', kanban_state_def)
         self.assertEqual(jira_issue_state, 'done')
-        jira_issue_state = states.transition(jira_issue_state, 'todo', kanban_states)
+        jira_issue_state = transition(jira_issue_state, 'todo', kanban_state_def)
         self.assertEqual(jira_issue_state, 'todo')
 
     def test_illegal_transition(self):
-        temperature_states = {
+        temperature_state_def = {
             'too hot': {
                 'next': ['just right'],
-                'prev': ['just right'],
             },
             'too cold': {
                 'next': ['just right'],
-                'prev': ['just right'],
             },
             'just right': {
                 'next': ['too hot', 'too cold'],
-                'prev': ['too hot', 'too cold'],
             }
         }
         room_state = 'too hot'
-        with self.assertRaises(states.TransitionException):
-            room_state = states.transition(room_state, 'too cold', temperature_states)
+        with self.assertRaises(TransitionException):
+            room_state = transition(room_state, 'too cold', temperature_state_def)
         self.assertEqual(room_state, 'too hot')
-        room_state = states.transition(room_state, 'just right', temperature_states)
-        room_state = states.transition(room_state, 'too cold', temperature_states)
+        room_state = transition(room_state, 'just right', temperature_state_def)
+        room_state = transition(room_state, 'too cold', temperature_state_def)
         self.assertEqual(room_state, 'too cold')
-        with self.assertRaises(states.TransitionException):
-            room_state = states.transition(room_state, 'too hot', temperature_states)
+        with self.assertRaises(TransitionException):
+            room_state = transition(room_state, 'too hot', temperature_state_def)
         self.assertEqual(room_state, 'too cold')
 
     def test_get_next_states(self):
-        temperature_states = {
+        temperature_state_def = {
             'too hot': {
                 'next': ['just right'],
-                'prev': ['just right'],
             },
             'too cold': {
                 'next': ['just right'],
-                'prev': ['just right'],
             },
             'just right': {
                 'next': ['too hot', 'too cold'],
-                'prev': ['too hot', 'too cold'],
             }
         }
         room_state = 'just right'
-        self.assertEqual(states.get_next_states(room_state, temperature_states), ['too hot', 'too cold'])
-        room_state = states.transition(room_state, 'too cold', temperature_states)
-        self.assertEqual(states.get_next_states(room_state, temperature_states), ['just right'])
-        room_state = states.transition(room_state, 'just right', temperature_states)
-        room_state = states.transition(room_state, 'too hot', temperature_states)
-        self.assertEqual(states.get_next_states(room_state, temperature_states), ['just right'])
-        self.assertNotEqual(states.get_next_states(room_state, temperature_states), ['just right', 'too cold'])
+        self.assertEqual(get_next_states(room_state, temperature_state_def), ['too hot', 'too cold'])
+        room_state = transition(room_state, 'too cold', temperature_state_def)
+        self.assertEqual(get_next_states(room_state, temperature_state_def), ['just right'])
+        room_state = transition(room_state, 'just right', temperature_state_def)
+        room_state = transition(room_state, 'too hot', temperature_state_def)
+        self.assertEqual(get_next_states(room_state, temperature_state_def), ['just right'])
+        self.assertNotEqual(get_next_states(room_state, temperature_state_def), ['just right', 'too cold'])
 
 
 if __name__ == '__main__':
